During testing for the 0.9.5 release, I ran into that familiar bug again (under Linux w/Python 3) where using message queue thread notification would raise interpreter errors like Invalid Thread State and the like.
In response to this, I changed to using the very straightforward code
recommended in the documentation which hasn't change since Python 2.5 --

	PyGILState_STATE gstate;
	gstate = PyGILState_Ensure();

	/* Perform Python actions here.  */
	result = CallSomeFunction();
	/* evaluate result */

	/* Release the thread. No Python API allowed beyond this point. */
	PyGILState_Release(gstate);

Why I felt compelled to add to this I don't remember, except that I think this
code just didn't work for me. My guess is that I was testing with Python 2.5
and the underlying code has improved in the Python 2.7/Python 3 that I'm
testing with now.

I've been unable to get this new, simple code to fail in my testing, which
makes it better than what I had before. I'm cautiously optimistic that it's
now correct.